Rename duration_serializer.rb to serializers.rb
